requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
twilio==8.5.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.0.0
pydantic-settings==2.0.3
starlette==0.28.0
redis==5.0.0
python-multipart==0.0.6
asyncpg==0.29.0
asyncpg
psycopg2-binary==2.9.9
boto3==1.33.13
websockets==11.0.3
